--- data/cache/APPLE_INC/docling/Apple_Environmental_Progress_Report_2024_Global-table-1.csv
@@ -1,28 +1,0 @@
-,,,Fiscal year.2023,Fiscal year.2022,Fiscal year.2021,Fiscal year.2020,Fiscal year.2019
-0,Corporate emissions  (metric tons CO2e) 2,Gross emissions,"324,100","324,000","166,380","334,430","573,730"
-1,Corporate emissions  (metric tons CO2e) 2,Scope 1,"55,200","55,200","55,200","47,430","52,730"
-2,Corporate emissions  (metric tons CO2e) 2,"Natural gas, diesel, propane","35,300","39,700","40,070","39,340","40,910"
-3,Corporate emissions  (metric tons CO2e) 2,Fleet vehicles,"17,000","12,600","12,090","4,270","6,950"
-4,Corporate emissions  (metric tons CO2e) 2,Other emissions 3,"2,900","2,900","3,040","3,830","4,870"
-5,Corporate emissions  (metric tons CO2e) 2,Scope 2 (market-based) 4,"3,400","3,000","2,780",0,0
-6,Corporate emissions  (metric tons CO2e) 2,Electricity,0,0,0,0,0
-7,Corporate emissions  (metric tons CO2e) 2,"Steam, heating, and cooling 5","3,400","3,000","2,780",0,0
-8,Corporate emissions  (metric tons CO2e) 2,Scope 3,"412,800","265,800","108,400","287,000","521,000"
-9,Corporate emissions  (metric tons CO2e) 2,Business travel,"225,700","113,500","22,850","153,000","326,000"
-10,Corporate emissions  (metric tons CO2e) 2,Employee commute 6,"164,100","134,200","85,570","134,000","195,000"
-11,Corporate emissions  (metric tons CO2e) 2,Upstream impacts (scope 1),"18,300","10,600",0,0,0
-12,Corporate emissions  (metric tons CO2e) 2,Work from home (market-based),"4,700","7,500",0,0,0
-13,Corporate emissions  (metric tons CO2e) 2,Transmission and distribution loss (market-based),N/A,N/A,N/A,N/A,N/A
-14,Corporate emissions  (metric tons CO2e) 2,Third-party cloud (market-based),0,0,0,0,0
-15,Corporate emissions  (metric tons CO2e) 2,Carbon removals,,,,,
-16,Corporate emissions  (metric tons CO2e) 2,Corporate carbon offsets 7,"-471,400","-324,100 8","-167,000 9","-70,000 10",0
-17,Product life  cycle emissions  (metric tons CO2e) 11,Gross emissions (Scope 3),"15,570,000","20,280,000","23,020,000","22,260,000","24,460,000"
-18,Product life  cycle emissions  (metric tons CO2e) 11,Manufacturing (purchased goods and services),"9,400,000","13,400,000","16,200,000","16,100,000","18,900,000"
-19,Product life  cycle emissions  (metric tons CO2e) 11,Product transportation (upstream and downstream),"1,500,000","1,900,000","1,750,000","1,800,000","1,400,000"
-20,Product life  cycle emissions  (metric tons CO2e) 11,Product use (use of sold products),"4,600,000","4,900,000","4,990,000","4,300,000","4,100,000"
-21,Product life  cycle emissions  (metric tons CO2e) 11,End-of-life processing,"70,000","80,000","80,000","60,000","60,000"
-22,,Carbon removals,,,,,
-23,,Product carbon offsets,"-13,500",0,"-500,000 12",0,0
-24,Total gross scope 3 emissions (corporate and product)  (metric tons CO2e),Total gross scope 3 emissions (corporate and product)  (metric tons CO2e),"15,980,000","20,545,800","23,128,400","22,550,000","24,980,000"
-25,Total gross carbon footprint (without offsets)  (metric tons CO2e) 13,Total gross carbon footprint (without offsets)  (metric tons CO2e) 13,"16,100,000","20,600,000","23,200,000","22,600,000","25,100,000"
-26,Total net carbon footprint (after applying offsets)  (metric tons CO2e) 12,Total net carbon footprint (after applying offsets)  (metric tons CO2e) 12,"15,600,000","20,300,000","22,530,000","22,530,000","25,100,000"
