--- data/cache/APPLE_INC/docling/Apple_Environmental_Progress_Report_2024_Global-table-2.csv
@@ -1,24 +1,0 @@
-,Electricity (million kWh),Renewable electricity  (million kWh),"Scope 2 emissions  (market-based, metric  tons CO2e) 1"
-0,940,940,0
-1,427,427,0
-2,14,14,0
-3,113,113,0
-4,143,143,0
-5,16,16,0
-6,18,18,0
-7,41,41,0
-8,168,168,0
-9,"2,344","2,344",0
-10,453,453,0
-11,488,488,0
-12,0,0,0
-13,269,269,0
-14,440,440,0
-15,40,40,0
-16,387,387,0
-17,96,96,0
-18,171,171,0
-19,203,203,0
-20,96,96,0
-21,107,107,0
-22,"3,199","3,199",0
