--- data/cache/APPLE_INC/docling/Apple_Environmental_Progress_Report_2024_Global-table-3.csv
@@ -1,24 +1,0 @@
-,Location,Scope 1.Total gas (MMBtu),Renewable.biogas (MMBtu),Scope 1 emissions (metric.tons CO2e)
-0,Corporate,"1,007,071","746,506","42,746"
-1,"Cupertino, CA","805,271","202,306","32,027"
-2,"Elk Grove, CA","10,175",0,540
-3,"Austin, TX","24,826",0,"1,319"
-4,Other U.S.,"104,718","544,200","5,562"
-5,"Cork, Ireland","24,219",0,"1,286"
-6,Singapore,518,0,28
-7,China,960,0,51
-8,Other international,"36,384",0,"1,933"
-9,Data center,740,0,39
-10,"Maiden, NC",0,0,0
-11,"Mesa, AZ",524,0,28
-12,"Newark, CA",0,0,0
-13,"Prineville, OR",216,0,11
-14,"Reno, NV",0,0,0
-15,"Viborg, Denmark",-,-,-
-16,Colocation facilities (U.S.),-,-,-
-17,Colocation facilities (international),-,-,-
-18,China,-,-,-
-19,Retail stores,"58,446",0,"3,105"
-20,Domestic (U.S.),"34,449",0,"1,830"
-21,International,"23,997",0,"1,275"
-22,Total,"1,066,257","746,506","45,890"
