--- data/cache/APPLE_INC/docling/Apple_Environmental_Progress_Report_2024_Global-table-4.csv
@@ -1,10 +1,0 @@
-,Parameter,Quantity,Units,Boundary / Protocol
-0,Natural Gas Consumption,"1,066,300",Metric million British  thermal unit,Worldwide occupied  properties / Apple Internal  Protocol
-1,Electricity Consumption,"3,500",Million kilowatt  hours (mkWh),Worldwide occupied  properties / Apple Internal  Protocol
-2,Renewable Energy,"3,500",Million kilowatt  hours (mkWh),Worldwide / Invoiced  quantities & self-generated
-3,Scope 1 GHG Emissions,"55,200",Metric tonnes of  carbon dioxide  equivalent (tCO2e),Worldwide occupied  properties / WRI/WBCSD  GHG Protocol
-4,Scope 2 GHG Emissions  (Location-Based),"1,206,700",tCO2e,Worldwide occupied  properties / WRI/WBCSD  GHG Protocol
-5,Scope 2 GHG Emissions  (Market-Based),"3,400",tCO2e,Worldwide occupied  properties / WRI/WBCSD  GHG Protocol
-6,Scope 3 Transmission and  Distribution Loses (Market- Based),0,tCO2e,Worldwide occupied  properties / WRI/WBCSD  GHG Protocol Value Chain  (Scope 3)
-7,Scope 3 GHG Emissions -  Upstream Fuel-Related  Activities,"18,300",tCO2e,Worldwide occupied  properties / WRI/WBCSD  GHG Protocol Value Chain  (Scope 3)
-8,Scope 3 GHG Emissions -  Business Travel,"225,700",tCO2e,Worldwide occupied  properties / WRI/WBCSD  GHG Protocol Value Chain  (Scope 3)
