--- data/cache/APPLE_INC/docling/Apple_Environmental_Progress_Report_2024_Global-table-5.csv
@@ -1,15 +1,0 @@
-,Scope 3 GHG Emissions -  Employee Commute,"164,100",tCO2e,Worldwide occupied  properties / WRI/WBCSD  GHG Protocol Value Chain  (Scope 3)
-0,Scope 3 GHG Emissions -  Work From Home Emissions  (Employee Commute)  (Location-Based),"20,600",tCO2e,Worldwide occupied  properties / WRI/WBCSD  GHG Protocol Value Chain  (Scope 3)
-1,Scope 3 GHG Emissions -  Work From Home Emissions  (Employee Commute)   (Market-Based),"4,700",tCO2e,Worldwide occupied  properties / WRI/WBCSD  GHG Protocol Value Chain  (Scope 3)
-2,Scope 3 GHG Emissions -  Other Cloud Services  (Purchased Goods and  Services) (Market-Based),0,tCO2e,Worldwide occupied  properties / WRI/WBCSD  GHG Protocol Value Chain  (Scope 3)
-3,Water Withdrawal,1600,Million gallons,Worldwide occupied  properties / Apple Internal  Protocol
-4,Water Discharge,900,Million gallons,Worldwide occupied  properties / Apple Internal  Protocol
-5,Trash Disposed in Landfill,"17,400",Metric tonnes,Worldwide occupied  properties / Apple Internal  Protocol
-6,Hazardous Waste  (Regulated waste),"3,300",Metric tonnes,Worldwide occupied  properties / Apple Internal  Protocol
-7,Recycled Material (Removal  by recycling contractor),"37,000",Metric tonnes,Worldwide occupied  properties / Apple Internal  Protocol
-8,Composted Material,"6,700",Metric tonnes,Worldwide occupied  properties / Apple Internal  Protocol
-9,Waste to Energy,"2,600",Metric tonnes  .,Worldwide occupied  properties / Apple Internal  Protocol
-10,C&D Landfilled,"3,200",Metric tonnes,Worldwide occupied  properties / Apple Internal  Protocol
-11,C&D Recycled,"23,600",Metric tonnes,Worldwide occupied  properties / Apple Internal  Protocol
-12,Paper Used,"1,100",Metric tonnes,Worldwide occupied  properties / Apple Internal  Protocol
-13,Product end use avoided  emissions,"48,800",tCO2e,Worldwide occupied  properties / WRI/WBCSD  GHG Protocol Value Chain  (Scope 3)
