--- data/cache/APPLE_INC/docling/Apple_Environmental_Progress_Report_2024_Global-table-6.csv
@@ -1,7 +1,0 @@
-,Comprehensive Carbon Footprint (Market Based),Comprehensive Carbon Footprint (Market Based),Comprehensive Carbon Footprint (Market Based),Comprehensive Carbon Footprint (Market Based)
-0,Corporate GHG Emissions  (Market-Based) 1,"471,400",tCO2e,Worldwide occupied  properties / WRI/WBCSD  GHG Protocol
-1,Product Use 2,"4,600,000",tCO2e,Worldwide occupied  properties / WRI/WBCSD  GHG Protocol
-2,Manufacturing 3,"9,400,000",tCO2e,Worldwide occupied  properties / WRI/WBCSD  GHG Protocol
-3,Transportation 4,"1,500,000",tCO2e,Worldwide occupied  properties / WRI/WBCSD  GHG Protocol
-4,Recycling 4,"70,000",tCO2e,Worldwide occupied  properties / WRI/WBCSD  GHG Protocol
-5,Comprehensive Carbon  Footprint 5,"16,100,000",tCO2e,Worldwide occupied  properties / WRI/WBCSD  GHG Protocol
