--- data/cache/APPLE_INC/docling/Apple_Environmental_Progress_Report_2024_Global-table-7.csv
@@ -1,3 +1,0 @@
-,Parameter,Quantity,Units,Boundary / Protocol
-0,"485,000",tCO2e,Worldwide occupied  properties / WRI/WBCSD  GHG Protocol Value Chain,Carbon Removals
-1,"15,600,000",tCO2e,Worldwide occupied  properties / WRI/WBCSD,Net Footprint 6
