+++ data/cache/BANK_OF_AMERICA_CORP/docling/Sustainability_at_Bank_of_America_2024_Report-table-1.csv
@@ -1,0 +1,6 @@
+,Organization/Initiative,Sector,Overview
+0,First Movers Coalition,Multiple,Mobilizes collective demand for critical emerging technologies essential for the Net Zero transition. The FMC  is a coalition of companies using their purchasing power to create early markets for innovative clean  technologies across seven hard-to-abate sectors.
+1,Airports of Tomorrow,Aviation,"Addresses the technology, policy and financing needs of the transition to Net Zero aviation through strategic  dialogues between public and private stakeholders across the aviation value chain."
+2,Sustainable Aviation Buyers  Alliance (SABA),Aviation,"Sustainable Aviation Buyers Alliance aims to accelerate the path to Net Zero aviation by driving investment in  and adoption of sustainable aviation fuel (SAF). Members work to develop a rigorous, transparent system that  expands opportunities to invest in high-integrity SAF to all businesses and organizations interested in reducing  the climate impacts of flying."
+3,Breakthrough Energy  Catalyst,Multiple,"Breakthrough Energy Catalyst (Catalyst), a program within the larger Breakthrough Energy network founded by  Bill Gates, is a novel platform that funds and invests in first-of-a-kind commercial projects for emerging  climate technologies. By investing in these opportunities, Catalyst seeks to accelerate the adoption of these  technologies worldwide and reduce their costs. Bank of America was one of the program's first private sector  partners when we invested in 2021."
+4,Mission Possible Partnership,Multiple,MPP is a movement of climate leaders and companies driving industrial decarbonization across the entire  value chain of the world's highest-emitting heavy industry and transport sectors.
