+++ data/cache/BANK_OF_AMERICA_CORP/docling/Sustainability_at_Bank_of_America_2024_Report-table-10.csv
@@ -1,0 +1,3 @@
+,Category,Units,Committed
+0,Physical Intensity (Auto Manufacturing) 1,g CO2e/km,215
+1,"Physical Intensity (Energy - Energy - Oil & Gas Refiners and  Producers includes Scope 1, Scope 2) 1",g CO2e/MJ,7
