+++ data/cache/BANK_OF_AMERICA_CORP/docling/Sustainability_at_Bank_of_America_2024_Report-table-3.csv
@@ -1,0 +1,17 @@
+,GHG Emissions by Region,Units,2023 Gross Location-Based Emissions.Scope  1 direct  emissions,2023 Gross Location-Based Emissions.Scope 2  indirect  emissions,2023 Gross Location-Based Emissions.Total  Scope 1  and  Scope 2  emissions,2023 Gross Market-Based Emissions.Scope  1 direct  emissions,2023 Gross Market-Based Emissions.Scope 2  indirect  emissions,2023 Gross Market-Based Emissions.Total  Scope 1  and  Scope 2  emissions
+0,U.S. and Canada,Metric tons CO 2 e,"64,091","510,430","574,521","64,091","4,842","68,934"
+1,Asia Pacific,Metric tons CO 2 e,598,"78,223","78,821",598,"12,246","12,844"
+2,"Europe, Middle East and Africa (EMEA)",Metric tons CO 2 e,"3,298","19,979","23,277","3,298",593,"3,891"
+3,Latin America,Metric tons CO 2 e,62,"1,381","1,443",62,55,117
+4,GHG Emissions by Country,,,,,,,
+5,United States,Metric tons CO 2 e,"63,978","510,311","574,289","63,978","4,842","68,821"
+6,India,Metric tons CO 2 e,267,"48,280","48,547",267,374,641
+7,United Kingdom,Metric tons CO 2 e,"2,303","16,777","19,080","2,303",0,"2,303"
+8,China,Metric tons CO 2 e,139,"9,905","10,044",139,208,347
+9,Singapore,Metric tons CO 2 e,13,"7,129","7,143",13,"7,129","7,143"
+10,Japan,Metric tons CO 2 e,62,"5,777","5,839",62,1,62
+11,Australia,Metric tons CO 2 e,1,"1,994","1,995",1,"1,060","1,061"
+12,South Korea,Metric tons CO 2 e,66,"1,769","1,834",66,"1,769","1,834"
+13,Taiwan,Metric tons CO 2 e,18,"1,685","1,703",18,"1,685","1,703"
+14,Ireland,Metric tons CO 2 e,553,729,"1,281",553,0,553
+15,Rest of World,Metric tons CO 2 e,651,"5,658","6,309",651,669,"1,320"
