+++ data/cache/BANK_OF_AMERICA_CORP/docling/Sustainability_at_Bank_of_America_2024_Report-table-4.csv
@@ -1,0 +1,11 @@
+,Carbon Credit Reporting,Units,2021,2022,2023
+0,Total carbon credits retired,Total carbon credits retired,Total carbon credits retired,Total carbon credits retired,Total carbon credits retired
+1,Total carbon credits retired,Metric tons CO 2 e,"89,786","164,741","178,605"
+2,Carbon credits retired by type of credit,Carbon credits retired by type of credit,Carbon credits retired by type of credit,Carbon credits retired by type of credit,Carbon credits retired by type of credit
+3,Avoidance carbon credits retired,Metric tons CO 2 e,"53,786","94,140","100,806"
+4,Removal carbon credits retired,Metric tons CO 2 e,"36,000","70,601","77,799"
+5,Total carbon credits retired,Metric tons CO 2 e,"89,786","164,741","178,605"
+6,Carbon credits retired by scope applied to,Carbon credits retired by scope applied to,Carbon credits retired by scope applied to,Carbon credits retired by scope applied to,Carbon credits retired by scope applied to
+7,Scope 1 and 2,Metric tons CO 2 e,"70,963","84,569","85,786"
+8,Scope 3 Business Travel,Metric tons CO 2 e,"18,823","80,172","92,819"
+9,Total carbon credits retired,Metric tons CO 2 e,"89,786","164,741","178,605"
