+++ data/cache/BANK_OF_AMERICA_CORP/docling/Sustainability_at_Bank_of_America_2024_Report-table-5.csv
@@ -1,0 +1,5 @@
+,Facilities,Units,2010  (baseline),2021,2022,2023
+0,LEED® (or comparable) certifications,Net square feet Percent of total  square footage,"12,537,553 10%","17,882,033 25%","18,398,417 26%","21,157,019  32%"
+1,Land use,Units,2010  (baseline),2021,2022,2023
+2,,Sites that  intersected with  areas protected for  biodiversity,Not available,10,13,13
+3,Land use and ecological sensitivity - U.S.,Area of buildings  (square meters),Not available,"7,120","7,500","7,500"
