+++ data/cache/BANK_OF_AMERICA_CORP/docling/Sustainability_at_Bank_of_America_2024_Report-table-6.csv
@@ -1,0 +1,29 @@
+,0,1,2,3,4,5,6,7,8,9
+0,"Emissions  Scope 1 GHG Emissions 68,050.00  tCO2e","Emissions  Scope 1 GHG Emissions 68,050.00  tCO2e","Emissions  Scope 1 GHG Emissions 68,050.00  tCO2e","Emissions  Scope 1 GHG Emissions 68,050.00  tCO2e","Emissions  Scope 1 GHG Emissions 68,050.00  tCO2e","Emissions  Scope 1 GHG Emissions 68,050.00  tCO2e","Emissions  Scope 1 GHG Emissions 68,050.00  tCO2e","Emissions  Scope 1 GHG Emissions 68,050.00  tCO2e","Emissions  Scope 1 GHG Emissions 68,050.00  tCO2e","Emissions  Scope 1 GHG Emissions 68,050.00  tCO2e"
+1,"Scope 2 GHG Location-Based Emissions 610,013.00  tCO2e","Scope 2 GHG Location-Based Emissions 610,013.00  tCO2e","Scope 2 GHG Location-Based Emissions 610,013.00  tCO2e",,,,,,,
+2,Scope 2 GHG Market-Based Emissions,Scope 2 GHG Market-Based Emissions,Scope 2 GHG Market-Based Emissions,"17,736.00  tCO2e","17,736.00  tCO2e",,,,,
+3,Scope 1 and Scope 2 Carbon Credits Retired,Scope 1 and Scope 2 Carbon Credits Retired,Scope 1 and Scope 2 Carbon Credits Retired,"85,786.00  tCO2e","85,786.00  tCO2e",,,,,
+4,,Purchased Goods and  Services,Purchased Goods and  Services,"1,722,654  tCO2e","1,722,654  tCO2e",,,,,
+5,,,,"48,570  tCO2e","48,570  tCO2e",,,,,
+6,,Fuel- and Energy-Related  Activities,Fuel- and Energy-Related  Activities,"168,018  tCO2e","168,018  tCO2e",,,,,
+7,,Upstream Transportation  and Distribution,Upstream Transportation  and Distribution,"152,752  tCO2e","152,752  tCO2e",,,,,
+8,,Waste (Traditional  Disposal),Waste (Traditional  Disposal),"18,406  tCO2e","18,406  tCO2e",,,,,
+9,,,,"92,818  tCO2e  92,819  tCO2e","92,818  tCO2e  92,819  tCO2e",,,,,
+10,,,,,,,,,,
+11,,Employee Commuting,Employee Commuting,-  tCO 2e,-  tCO 2e,,,,,
+12,,,,,,,,,,
+13,,,Scope 1  direct emissions (tCO2e),Scope 2  indirect emissions -  location-based (tCO2e),Scope 2  indirect emissions  market-based (tCO2e),Total Scope 1 and Scope  2 emissions - location- based (tCO2e),Total Scope 1 and Scope  2 emissions - market- based (tCO2e),Total Scope 1 and Scope  2 emissions - market- based (tCO2e),Total Scope 1 and Scope  2 emissions - market- based (tCO2e),Total Scope 1 and Scope  2 emissions - market- based (tCO2e)
+14,U.S. and Canada,,,"64,091","510,430.00",,,"4,842.00","574,521.00","68,934.00"
+15,Asia Pacific,,,598,"78,223.00",,"12,844.00","12,844.00","12,844.00","12,844.00"
+16,EMEA,,,"3,298","19,979.00",,,593,"23,277.00","3,891.00"
+17,Carbon Credits,Carbon Credits,Carbon Credits,,,,,,,
+18,Avoidance Carbon Credits Retired,,,"100,806  tCO2e",,,,,,
+19,,Removal Carbon Credits Retired,Removal Carbon Credits Retired,"77,799  tCO2e",,,,,,
+20,,Total Carbon Credits Retired,Total Carbon Credits Retired,"178,605  tCO2e","178,605  tCO2e",,,,,
+21,,Emissions of Ozone-Depleting Substances,Emissions of Ozone-Depleting Substances,,,,,,,
+22,,,,2  Metric tons,2  Metric tons,,,,,
+23,,,,19  Metric tons  2  Metric tons,19  Metric tons  2  Metric tons,,,,,
+24,Nitrous Oxides (NOX),,,,,,,,,
+25,,Sulfur Oxides (SOX),,,,,,,,
+26,,,,,,,,,,
+27,,Volatile Organic Compounds (VOCs),,,,,,,,
