+++ data/cache/BANK_OF_AMERICA_CORP/docling/Sustainability_at_Bank_of_America_2024_Report-table-7.csv
@@ -1,0 +1,11 @@
+,Category,Units,Utilized,Data  Quality
+0,"Absolute Total Financed Emissions (Auto Manufacturing includes Scope 1, Scope 2, and Scope 3)",Thousand Metric tons CO2e,"1,208",2.0 1  2.0 2
+1,Absolute Total Financed Emissions (Aviation includes Scope 1),Thousand Metric tons CO2e,"4,152",2.5
+2,Absolute Total Financed Emissions (Cement includes Scope 1 and Scope 2),Thousand Metric tons CO2e,"1,779",2.2
+3,"Absolute Total Financed Emissions (Energy - Pureplay thermal coal extraction includes Scope 1, Scope 2 and Scope 3)",Thousand Metric tons CO2e,339,5.0
+4,"Absolute Total Financed Emissions (Energy - Upstream producers, refiners, and integrated companies in the oil and gas  industry - includes Scope 1 and Scope 2)",Thousand Metric tons CO2e,"1,961",2.7
+5,"Absolute Total Financed Emissions (Energy -- Upstream producers, refiners, and integrated companies in the oil and gas  industry - includes Scope 3.11)",Thousand Metric tons CO2e,"18,593",3.4
+6,Absolute Total Financed Emissions (Energy - Midstream and downstream oil and gas companies - includes Scope 1 and  Scope 2),Thousand Metric tons CO2e,"2,605",3.6
+7,Absolute Total Financed Emissions (Power Generation - Scope 1 only),Thousand Metric tons CO2e,"3,496",2.9
+8,Absolute Total Financed Emissions (Iron & Steel includes Scope 1 and Scope 2),Thousand Metric tons CO2e,"2,708",2.6
+9,Portfolio Alignment Score - (Maritime Shipping Asset Specific includes Scope 1),Percent,-0.47,2.0
