+++ data/cache/BANK_OF_AMERICA_CORP/docling/Sustainability_at_Bank_of_America_2024_Report-table-8.csv
@@ -1,0 +1,17 @@
+,Category,Units,Utilized
+0,"Economic Intensity (Auto Manufacturing includes Scope 1, Scope 2, and Scope 3)",Metric tons CO2e/ Million USD  exposure,"1,196"
+1,Economic Intensity (Aviation includes Scope 1),Metric tons CO2e/ Million USD,921
+2,Economic Intensity (Cement includes Scope 1 and Scope 2),exposure  Metric tons CO2e/ Million USD,"2,847"
+3,"Economic Intensity (Energy - Oil & Gas Refiners and Producers includes Scope 1, Scope 2 and Scope 3)",exposure  Metric tons CO2e/ Million USD,"3,670"
+4,Economic Intensity (Energy - Midstream and downstream oil and gas companies - includes Scope 1 and Scope 2),exposure  Metric tons CO2e/ Million USD  exposure,448
+5,"Economic Intensity (Energy - Pureplay thermal coal extraction includes Scope 1, Scope 2 and Scope 3)",Metric tons CO2e/ Million USD  exposure,"21,787"
+6,Economic Intensity (Power Generation - Scope 1 only),Metric tons CO2e/ Million USD  exposure,371
+7,Economic Intensity (Iron & Steel includes Scope 1 and Scope 2),Metric tons CO2e/ Million USD,"1,810"
+8,Category,exposure  Units,Utilized
+9,"Physical Intensity (Auto Manufacturing includes Scope 1, Scope 2, and Scope 3)",g CO2e/km,201.6
+10,Physical Intensity (Aviation includes Scope 1),gCO2e/RTK,940.1
+11,Physical Intensity (Cement includes Scope 1 and Scope 2),tCO2e/tCP,0.651
+12,"Physical Intensity (Energy - Energy - Oil & Gas Refiners and Producers includes Scope 1, Scope 2)",g CO2e/MJ,6.5
+13,Physical Intensity (Energy - Oil & Gas Refiners and Producers includes Scope 3),g CO2/MJ,59.7
+14,Physical Intensity (Power Generation - Scope 1 only),kg CO2/MWh,322.2
+15,Physical Intensity (Iron & Steel includes Scope 1 and Scope 2),Metric tons CO2e/Metric Tons CS,1.77
