+++ data/cache/BANK_OF_AMERICA_CORP/docling/Sustainability_at_Bank_of_America_2024_Report-table-9.csv
@@ -1,0 +1,3 @@
+,Category,Units,Utilized
+0,Absolute Total Financed Emissions (Auto Manufacturing) 1,Thousand Metric  tons CO2e,"1,011"
+1,Absolute Total Financed Emissions (Aviation includes Scope  1) 1,Thousand Metric  tons CO2e,"3,486"
