+++ data/cache/std_esg_data.csv
@@ -1,0 +1,3 @@
+Metric,2023,2022,2021,2020,2019,Units
+Scope 1,"55,200","55,200","55,200","47,430","52,730",metric tons CO2e
+Scope 2 (market-based),"3,400","3,000","2,780",0,0,metric tons CO2e
