+++ data/cache/testing_apple.csv
@@ -1,0 +1,12 @@
+,GHG Emissions,Units,2010 (baseline),2021,2022,2023
+0,Scope 1 Direct Emissions,Metric tons CO2e,"106,870","57,076","66,775","68,050"
+1,Location-Based Scope 2 Indirect Emissions,Metric tons CO2e,"1,678,547","601,906","634,510","610,013"
+2,Total Scope 1 and Location-Based Scope 2 Emissions,Metric tons CO2e,"1,785,417","658,982","701,285","678,063"
+3,Reduction in Total Scope 1 and Location-Based Scope 2 Emissions,Metric tons CO2e,Not applicable,63%,61%,62%
+4,Scope 1 and Market-Based Scope 2 Emissions,Metric tons CO2e,Scope 1 and Market-Based Scope 2 Emissions,Scope 1 and Market-Based Scope 2 Emissions,Scope 1 and Market-Based Scope 2 Emissions,Scope 1 and Market-Based Scope 2 Emissions
+5,Scope 1 Direct Emissions,Metric tons CO2e,"106,870","57,076","66,775","68,050"
+6,Market-Based Scope 2 Indirect Emissions,Metric tons CO2e,"1,644,068","13,886","17,794","17,736"
+7,Total Gross Scope 1 and Market-Based Scope 2 Emissions,Metric tons CO2e,"1,750,939","70,963","84,569","85,786"
+8,Total Net Scope 1 and Market-Based Scope 2 Emissions,Metric tons CO2e,"1,750,939",0,0,0
+9,Reduction in Total Net Scope 1 and Market-Based Scope 2 Emissions,Metric tons CO2e,Not applicable,100%,100%,100%
+10,Metric tons CO 2 e,"70,963","84,569","85,786",,
