+++ src/extract/filtered_data.py
@@ -11,7 +11,7 @@ sys.path.append(os.getenv("ROOT_DIR"))
 
 from src.utils.units import get_units_raw_input, infer_units_for_rows  # noqa: E402
 from src.utils.data_models import TableParsers  # noqa: E402
-from src.utils.standardize_table_rework import standardize_table  # noqa: E402
+from src.utils.standardize_table import standardize_table # noqa: E402
 
 
 class Filter(BaseModel):
@@ -131,7 +131,7 @@ class Filter(BaseModel):
 if __name__ == "__main__":
     ROOT_DIR = os.getenv("ROOT_OUTPUT_PATH")
     filter_obj = Filter(
-        directory_path=os.path.join(ROOT_DIR, "APPLE_INC"),
+        directory_path=os.path.join(ROOT_DIR, "BANK_OF_AMERICA_CORP"),
         parser=TableParsers.DOCLING,
     )
     filter_obj.extract_filtered_df()
