+++ src/scripts/retrieve_emissions_data.py
@@ -133,7 +133,7 @@ def get_emissions_data(identifier, idType, parser):
 if __name__ == "__main__":
     start = time.time()
 
-    identifier = "US0378331005"
+    identifier = "US0605051046"
     idType = "isin"
     parser = TableParsers.DOCLING
     data = get_emissions_data(identifier, idType, parser)
